+++ packages/forms/src/FormField.tsx
@@ -1,4 +1,7 @@
+import { FormFieldCheckbox } from './FormFieldCheckbox';
 import { FormFieldInput } from './FormFieldInput';
+import { FormFieldRadio } from './FormFieldRadio';
+import { FormFieldSelect } from './FormFieldSelect';
 
 const FormField = () => {
   // to be implemented
@@ -6,5 +9,9 @@ const FormField = () => {
 };
 
 FormField.Input = FormFieldInput;
+FormField.Radio = FormFieldRadio;
+FormField.Select = FormFieldSelect;
+
+FormField.Checkbox = FormFieldCheckbox;
 
 export { FormField };